Move realex endpoints to separate route
