+++ packages/building-blocks-sdk/services/payments/index.ts
@@ -190,10 +190,10 @@ export class Payments {
   }
 
   async getRealexPaymentObject(
-    query: paths["/api/v1/transactions/realex/paymentObject"]["get"]["parameters"]["query"],
+    query: paths["/api/v1/realex/paymentObject"]["get"]["parameters"]["query"],
   ) {
     return formatQueryResult(
-      this.client.GET("/api/v1/transactions/realex/paymentObject", {
+      this.client.GET("/api/v1/realex/paymentObject", {
         params: {
           query,
         },
